Move getObjectStyle
